pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "specterqa"
version = "0.3.0"
description = "AI persona-based behavioral testing for web apps. No test scripts. YAML-configured. Vision-powered."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "SyncTek LLC" },
]
keywords = ["testing", "qa", "ai", "behavioral", "persona", "vision", "playwright"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "anthropic>=0.39.0",
    "playwright>=1.48.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
native = [
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
]
mcp = [
    "mcp>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "build>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/SyncTek-LLC/specterqa"
Repository = "https://github.com/SyncTek-LLC/specterqa"
Issues = "https://github.com/SyncTek-LLC/specterqa/issues"

[project.scripts]
specterqa = "specterqa.cli.app:app"
specterqa-mcp = "specterqa.mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/specterqa"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires running app and API key",
    "dogfood: SpecterQA testing itself",
]
